+++ CHANGELOG.md
@@ -4,11 +4,14 @@ This page lists the main changes made to Myokit in each release.
 
 ## Unreleased
 - Added
+  - [#1088](https://github.com/myokit/myokit/pull/1088) The DataLogViewer now displays meta data for myokit DataLogs.
+  - [#1088](https://github.com/myokit/myokit/pull/1088) The HEKA file format now supports more meta data, including filters.
 - Changed
 - Deprecated
 - Removed
 - Fixed
   - [#1083](https://github.com/myokit/myokit/pull/1083) Fixed bug when passing initial state `Expression` objects to `myokit.step`.
+  - [#1088](https://github.com/myokit/myokit/pull/1088) The HEKA file format meta data now shows the correct value for fast capacitance correction.
 
 ## [1.37.0] - 2024-06-17
 - Added